Move viewStubDataBinding() because it isn't extension.
